Enabled float values of the frequency.
git-svn-id: svn://orca.physics.unc.edu/Orca@2454 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Serial/MITPulser/ORMITPulserModel.h
@@ -31,15 +31,15 @@
 		int				clockSpeed;
 		int				resistance;
 		int				dutyCycle;
-		int				frequency;
+		float			frequency;
 }
 
 #pragma mark ***Initialization
 - (void) dealloc;
 
 #pragma mark ***Accessors
-- (int) frequency;
-- (void) setFrequency:(int)aFrequency;
+- (float) frequency;
+- (void) setFrequency:(float)aFrequency;
 - (int) dutyCycle;
 - (void) setDutyCycle:(int)aDutyCycle;
 - (int) resistance;
